# openapi.yaml
openapi: 3.1.0
info:
  title: AI Chatbot Backend
  description: '

    Multi-provider AI chatbot and stateless workflow API.


    ## Authentication

    No authentication is required. API keys are server-side only.


    ## Rate Limiting

    Chat endpoints are rate-limited per client IP. Exceeding the limit returns **429**

    with a `Retry-After` header indicating when the client may retry.


    ## Request Tracing

    Every response includes an `x-request-id` header (server-generated UUID hex).

    Include this value when reporting errors.


    ## Server-Sent Events (SSE)

    Streaming endpoints return `text/event-stream`. Each event has the shape:

    ```

    event: <event-type>

    data: <JSON object>

    ```

    See individual endpoint descriptions for the event types they emit.

    '
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Local development
- url: http://localhost:8001
  description: Local development (run.sh)
paths:
  /v1/workflow:
    post:
      tags:
      - Workflow
      summary: Chat Workflow
      operationId: chat_workflow_v1_workflow_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/workflow/stream:
    post:
      tags:
      - Workflow
      summary: Stream Workflow
      operationId: stream_workflow_v1_workflow_stream_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/workflow/schema/stream:
    post:
      tags:
      - Workflow
      summary: Stream Schema Workflow
      operationId: stream_schema_workflow_v1_workflow_schema_stream_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /healthz:
    get:
      tags:
      - Health
      summary: Healthz
      operationId: healthz_healthz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /v1/model:
    get:
      tags:
      - Models
      summary: Get Model
      operationId: get_model_v1_model_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResponse'
  /v1/models:
    get:
      tags:
      - Models
      summary: List Models
      operationId: list_models_v1_models_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsResponse'
  /v1/providers:
    get:
      tags:
      - Models
      summary: List Providers
      operationId: list_providers_v1_providers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvidersResponse'
  /v1/providers/{provider}/default-model:
    patch:
      tags:
      - Models
      summary: Update Provider Default Model
      operationId: update_provider_default_model_v1_providers__provider__default_model_patch
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DefaultModelUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sessions/{session_id}:
    get:
      tags:
      - Sessions
      summary: Get Session Info
      operationId: get_session_info_v1_sessions__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Sessions
      summary: Update Session
      operationId: update_session_v1_sessions__session_id__patch
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Sessions
      summary: Delete Session Endpoint
      operationId: delete_session_endpoint_v1_sessions__session_id__delete
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/chat:
    post:
      tags:
      - Chat
      summary: Chat
      operationId: chat_v1_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/chat/stream:
    post:
      tags:
      - Chat
      summary: Chat Stream
      operationId: chat_stream_v1_chat_stream_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ChatRequest:
      properties:
        message:
          type: string
          maxLength: 100
          minLength: 1
          title: Message
        session_id:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Session Id
        provider:
          anyOf:
          - type: string
            maxLength: 32
            minLength: 1
          - type: 'null'
          title: Provider
        model:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Model
      type: object
      required:
      - message
      title: ChatRequest
    ChatResponse:
      properties:
        session_id:
          type: string
          title: Session Id
        reply:
          type: string
          title: Reply
        provider:
          type: string
          title: Provider
        model:
          type: string
          title: Model
      type: object
      required:
      - session_id
      - reply
      - provider
      - model
      title: ChatResponse
    DefaultModelUpdate:
      properties:
        model:
          type: string
          maxLength: 64
          minLength: 1
          title: Model
      type: object
      required:
      - model
      title: DefaultModelUpdate
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
        request_id:
          type: string
          title: Request Id
      type: object
      required:
      - detail
      - request_id
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HealthResponse:
      properties:
        status:
          type: string
          title: Status
          default: ok
        service:
          type: string
          title: Service
          default: ai-chatbot-backend
      type: object
      title: HealthResponse
    ModelResponse:
      properties:
        provider:
          type: string
          title: Provider
        model:
          type: string
          title: Model
      type: object
      required:
      - provider
      - model
      title: ModelResponse
    ModelsResponse:
      properties:
        models:
          items:
            $ref: '#/components/schemas/OpenRouterModel'
          type: array
          title: Models
      type: object
      required:
      - models
      title: ModelsResponse
    OpenRouterModel:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        created:
          type: integer
          title: Created
        description:
          type: string
          title: Description
      type: object
      required:
      - id
      - name
      - created
      - description
      title: OpenRouterModel
    ProviderInfo:
      properties:
        provider:
          type: string
          title: Provider
        default_model:
          type: string
          title: Default Model
      type: object
      required:
      - provider
      - default_model
      title: ProviderInfo
    ProvidersResponse:
      properties:
        default_provider:
          type: string
          title: Default Provider
        default_model:
          type: string
          title: Default Model
        available_providers:
          items:
            $ref: '#/components/schemas/ProviderInfo'
          type: array
          title: Available Providers
      type: object
      required:
      - default_provider
      - default_model
      - available_providers
      title: ProvidersResponse
    SchemaRequest:
      properties:
        query:
          type: string
          maxLength: 500
          minLength: 1
          title: Query
        workspace:
          type: string
          maxLength: 50
          title: Workspace
          default: logistics-schema
        provider:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Provider
        model:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Model
      type: object
      required:
      - query
      title: SchemaRequest
    SessionResponse:
      properties:
        session_id:
          type: string
          title: Session Id
        provider:
          type: string
          title: Provider
        model:
          type: string
          title: Model
      type: object
      required:
      - session_id
      - provider
      - model
      title: SessionResponse
    SessionUpdate:
      properties:
        provider:
          anyOf:
          - type: string
            maxLength: 32
            minLength: 1
          - type: 'null'
          title: Provider
        model:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Model
      type: object
      title: SessionUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WorkflowRequest:
      properties:
        query:
          type: string
          maxLength: 500
          minLength: 1
          title: Query
        workspace:
          type: string
          maxLength: 50
          title: Workspace
          default: general
        provider:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Provider
        model:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Model
      type: object
      required:
      - query
      title: WorkflowRequest
    WorkflowResponse:
      properties:
        result:
          type: string
          title: Result
        provider:
          type: string
          title: Provider
        model:
          type: string
          title: Model
      type: object
      required:
      - result
      - provider
      - model
      title: WorkflowResponse